refactor(app-service): refactor route mapping;
