+++ packages/app-service/src/router/function-debug.ts
@@ -7,11 +7,11 @@
 
 import { Request, Response } from 'express'
 import { FunctionContext } from 'cloud-function-engine'
-import { parseToken } from '../../lib/utils/token'
-import { logger } from '../../lib/logger'
-import { addFunctionLog } from '../../api/function'
+import { parseToken } from '../lib/utils/token'
+import { logger } from '../lib/logger'
+import { addFunctionLog } from '../api/function'
 import { ObjectId } from 'bson'
-import { CloudFunction } from '../../lib/function'
+import { CloudFunction } from '../lib/function'
 
 /**
  * Handler of debugging cloud function
